* Add option to delay Auto-Type in Password Change form
On some *nix systems, KeePass won't fall to background properly
after clicking one of the Auto-Type buttons in the Password Change form.
To compensate that, it's possible to add an initial delay before starting the Auto-Type.
This allows to manually switch to the target window and place the cursor in the proper field.

Closes #16